+++ v1/packages/assets/src/mainnet/cosmoshub.ts
@@ -2056,6 +2056,42 @@ const assets: AssetList = {
         }]
     },
     {
+      description: 'Dungeon native token',
+      denom_units: [{
+          denom: 'ibc/C57D2C02129637EE06447A85BDF35FD84C0DF75D7493840F152889CD1ACBA90E',
+          exponent: 0,
+          aliases: ['udgn']
+        }, {
+          denom: 'DGN',
+          exponent: 6
+        }],
+      base: 'ibc/C57D2C02129637EE06447A85BDF35FD84C0DF75D7493840F152889CD1ACBA90E',
+      name: 'Dungeon Coin',
+      display: 'DGN',
+      symbol: 'DGN',
+      logo_URIs: {
+        png: 'https://raw.githubusercontent.com/cosmos/chain-registry/master/dungeon/images/DGN.png'
+      },
+      images: [{
+          png: 'https://raw.githubusercontent.com/cosmos/chain-registry/master/dungeon/images/DGN.png',
+          theme: {
+            circle: true
+          }
+        }],
+      type_asset: 'sdk.coin',
+      traces: [{
+          type: 'ibc',
+          counterparty: {
+            channel_id: 'channel-3',
+            base_denom: 'udgn',
+            chain_name: 'dungeon'
+          },
+          chain: {
+            channel_id: 'channel-1213'
+          }
+        }]
+    },
+    {
       description: 'The native governance and staking token of the Dymension Hub',
       denom_units: [{
           denom: 'ibc/49ACCECDED4A6A2B5DD683FA048235F4F66B65F58CD558C9A441195502191D01',
@@ -7855,6 +7891,42 @@ const assets: AssetList = {
           },
           chain: {
             channel_id: 'channel-750'
+          }
+        }]
+    },
+    {
+      description: 'Dungeon native token',
+      denom_units: [{
+          denom: 'ibc/C57D2C02129637EE06447A85BDF35FD84C0DF75D7493840F152889CD1ACBA90E',
+          exponent: 0,
+          aliases: ['udgn']
+        }, {
+          denom: 'DGN',
+          exponent: 6
+        }],
+      base: 'ibc/C57D2C02129637EE06447A85BDF35FD84C0DF75D7493840F152889CD1ACBA90E',
+      name: 'Dungeon Coin',
+      display: 'DGN',
+      symbol: 'DGN',
+      logo_URIs: {
+        png: 'https://raw.githubusercontent.com/cosmos/chain-registry/master/dungeon/images/DGN.png'
+      },
+      images: [{
+          png: 'https://raw.githubusercontent.com/cosmos/chain-registry/master/dungeon/images/DGN.png',
+          theme: {
+            circle: true
+          }
+        }],
+      type_asset: 'sdk.coin',
+      traces: [{
+          type: 'ibc',
+          counterparty: {
+            channel_id: 'channel-3',
+            base_denom: 'udgn',
+            chain_name: 'dungeon'
+          },
+          chain: {
+            channel_id: 'channel-1213'
           }
         }]
     },