+++ v2/packages/cosmostation/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.70.9 (2024-11-11)
+
+**Note:** Version bump only for package @chain-registry/v2-cosmostation
+
+
+
+
+
+## 1.70.8 (2024-11-09)
+
+**Note:** Version bump only for package @chain-registry/v2-cosmostation
+
+
+
+
+
 ## 1.70.7 (2024-11-08)
 
 **Note:** Version bump only for package @chain-registry/v2-cosmostation