+++ v2/packages/keplr/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.29 (2024-11-11)
+
+**Note:** Version bump only for package @chain-registry/v2-keplr
+
+
+
+
+
+## 0.0.28 (2024-11-10)
+
+**Note:** Version bump only for package @chain-registry/v2-keplr
+
+
+
+
+
+## 0.0.27 (2024-11-09)
+
+**Note:** Version bump only for package @chain-registry/v2-keplr
+
+
+
+
+
 ## 0.0.26 (2024-11-08)
 
 **Note:** Version bump only for package @chain-registry/v2-keplr