+++ v2/workflows/cli/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.53.9 (2024-11-11)
+
+**Note:** Version bump only for package @chain-registry/cli
+
+
+
+
+
+## 1.53.8 (2024-11-09)
+
+**Note:** Version bump only for package @chain-registry/cli
+
+
+
+
+
 ## 1.53.7 (2024-11-08)
 
 **Note:** Version bump only for package @chain-registry/cli